autotest: make fuzzing suite portable
